Add more tests for vector element lifetimes

+++ src/debug.cpp
@@ -8,14 +8,14 @@ int main() {
     myVec.push_back(8);
 
     std::cout << "size: " << myVec.size() << " | capacity: " << myVec.capacity() << '\n';
-    for (int i{}; i < myVec.size(); ++i) {
+    for (size_t i{}; i < myVec.size(); ++i) {
         std::cout << myVec[i] << '\n';
         // std::cout << i << '\n';
     }
 
     myVec.pop_back();
 
-    for (int i{}; i < myVec.size(); ++i) {
+    for (size_t i{}; i < myVec.size(); ++i) {
         std::cout << myVec[i] << '\n';
     }
